+++ dash_json_grid/__init__.py
@@ -29,8 +29,9 @@ import dash as _dash
 # noinspection PyUnresolvedReferences
 from ._imports_ import DashJsonGrid as _DashJsonGrid
 from .mixins import MixinDataRoute as _MixinDataRoute, MixinFile as _MixinFile
+from .typehints import ThemeConfigs
 
-__all__ = ("DashJsonGrid",)
+__all__ = ("DashJsonGrid", "ThemeConfigs")
 
 if not hasattr(_dash, "__plotly_dash") and not hasattr(_dash, "development"):
     print(